Utilities folder was deleted, CanchaCard was named as courts and moved to the pages folder, Unused files were deleted
