+++ src/components/pages/Home.jsx
@@ -1,14 +1,40 @@
+import { useState } from 'react';
+import { useEffect } from 'react';
 import Container from "react-bootstrap/Container"
 import Row from "react-bootstrap/Row"
 import Col from 'react-bootstrap/Col'
 import Image from 'react-bootstrap/Image';
 import Carousel from 'react-bootstrap/Carousel';
 import Button from 'react-bootstrap/Button'
-import { Link } from 'react-router-dom'
-import { useState } from 'react';
 import Modal from 'react-bootstrap/Modal';
+import { Link } from 'react-router-dom'
+import Card from '../layout/Card'
 
 const Home = () => {
+
+  const [products, setProduct] = useState([])
+
+  const getProduct = async () => {
+    const response = await fetch ("http://localHost:4000/admin/getProducts",{
+        method:'GET',
+        headers:{'Content-type':'application/json'},
+        credentials:'include'
+    })
+    const responseData = await response.json();
+
+    const mappedProduct = responseData.map(product => ({
+      Title: product.Title,
+      description: product.description,
+      Url:product.Url
+    }))
+    setProduct(mappedProduct);
+    }
+    useEffect(() => {
+      getProduct();
+    }, []);
+    useEffect(() => {
+      console.log(products)
+    }, [products])
 
   const [showModal, setShowModal] = useState(false);
 
@@ -56,39 +82,34 @@ const Home = () => {
           </Col>
         </Row>
       </Container>
-      {/* <Container fluid className="my-3">
-        <div>
-          <h4 className="text-light">Alguno de nuestros productos...</h4>
-        </div>
+      <Container className='mb-4' fluid>
+        <h3 className='text-light mb-3 mx-3'>Esto y mucho mas vas a poder conseguir en el predio!</h3>
         <Row>
-          <CardProductos title={'titulo 1'} url={'https://picsum.photos/id/237/200'}/>
-          <CardProductos title={'titulo 2'} url={'https://picsum.photos/id/238/200'}/>
-          <CardProductos title={'titulo 3'} url={'https://picsum.photos/id/239/200'}/>
-          <CardProductos title={'titulo 4'} url={'https://picsum.photos/id/236/200'}/>
-          <CardProductos title={'titulo 5'} url={'https://picsum.photos/id/235/200'}/>
-          <CardProductos title={'titulo 6'} url={'https://picsum.photos/id/234/200'}/>
+          {
+            products.map((product, index) => (
+              <Card key={index} title={product.Title}
+              url={product.Url}/>
+            ))
+          }
         </Row>
-      </Container> */}
-      <Modal
-      show={showModal}
-      onHide={handleCloseModal}
-      >
-      <Modal.Header closeButton>
-        <Modal.Title>
-          Debes loguearte para alquilar!
-          </Modal.Title>
-      </Modal.Header>
-      <Modal.Body>
-        A continuacion logueate!
-      </Modal.Body>
-      <Modal.Footer>
-        <Link to='/login'>
-          <Button className='btn-login1' onClick={handleCloseModal}>
-            Login
-          </Button>
-        </Link>
-      </Modal.Footer>
-    </Modal>
+      </Container>
+      <Modal show={showModal} onHide={handleCloseModal}>
+        <Modal.Header closeButton>
+          <Modal.Title>
+            Debes loguearte para alquilar!
+            </Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          A continuacion logueate!
+        </Modal.Body>
+        <Modal.Footer>
+          <Link to='/login'>
+            <Button className='btn-login1' onClick={handleCloseModal}>
+              Login
+            </Button>
+          </Link>
+        </Modal.Footer>
+      </Modal>
   </>
   )
 }
