+++ src/components/routers/PrivateRoutes.jsx
@@ -5,9 +5,8 @@ import Home from "../pages/Home";
 import Nosotros from "../pages/Sobrenosotros";
 import Contacto from "../pages/Contacto";
 import Bookin from '../layout/Bookin'
-import Productos from '../utilities/ProductCard'
-import Canchas from '../utilities/Canchacard'
-import ProductCard from "../utilities/ProductCard";
+import Canchas from '../pages/Canchas'
+
 
 const PrivateRoutes = ({user}) => {
 
@@ -21,10 +20,10 @@ const PrivateRoutes = ({user}) => {
         <Route exact path='/contacto' element={<Contacto/>}/>
         <Route exact path='/aboutUs' element={<Nosotros/>}/>
         <Route exact path='/galeria' element={<Galeria/>}/>
-        <Route exact path='/productos' element={<Productos/>}/>
         <Route exact path='/alquiler' element={<Bookin/>}/>
         <Route exact path='/canchas' element={<Canchas/>}/>
         <Route exact path='/admin' element={<Admin/>}/>
+        
         <Route path='*' element={<Navigate to='/'/>}/>
       </Routes>
     </>
